services/requirements.txt
connexion[swagger-ui] >= 2.6.0
python_dateutil >= 2.6.0
setuptools >= 21.0.0
Flask-SQLAlchemy >= 2.4.4
GitPython >= 3.1.0
certifi >= 14.05.14
six >= 1.10
urllib3 >= 1.15.1
conan >= 1.28.0
